refactor backend classes into separate modules
