+++ bambi/__init__.py
@@ -1,3 +1,13 @@
 from .models import Model
 from .priors import Prior, Family
 from .results import ModelResults
+from .backends import StanBackEnd, PyMC3BackEnd
+
+__all__ = [
+    'Model',
+    'Prior',
+    'Family',
+    'ModelResults',
+    'StanBackEnd',
+    'PyMC3BackEnd'
+]
