+++ bambi/backends/__init__.py
@@ -1,0 +1,8 @@
+from .pymc import PyMC3BackEnd
+from .stan import StanBackEnd
+
+
+__all__ = [
+    'StanBackEnd',
+    'PyMC3BackEnd'
+]
