+++ bambi/backends/base.py
@@ -1,0 +1,17 @@
+from abc import ABCMeta, abstractmethod
+
+
+class BackEnd(object):
+
+    '''
+    Base class for BackEnd hierarchy.
+    '''
+    __metaclass__ = ABCMeta
+
+    @abstractmethod
+    def build(self):
+        pass
+
+    @abstractmethod
+    def run(self):
+        pass
